Create TicTacToe with rust
This creates a TicTacToe game. It is done as an exercise to learn rust.
It is a very basic command line game that has players alternate taking
turns by entering the position they would like to play (1-9). Player 1
is always 'x' and goes first. There is a small help menu accessed with
'help' and the game can be exited with 'quit'.