Reorganize game code into a library
This takes the one file of code and reorganizes it into a library with
modules. Each module has its own file so its not just one huge monolith.
The separation into a library and a binary that uses the library was
probably not necessary unless someone whats to add an easter egg of a
tictactoe game into their rust command line tool or something.