Add TUI display
This adds two new dependencies to the project, TUI and Termion. These
are used to create a GUI for the game. The GUI is almost identical to
the CLI version. Numbers are still required to put your mark and you can
still see the history layed out. Future work will hoepfully let users
move cursors to select tiles.